Addressed question about breadwinning status at birth
It should be status at durmom==0, or in the year the woman becomes a mother. Here we have a higher estimate of breadwinning than when we use 2008 data that allows us to estimate breadwinning in the 4 months during which she had a birth. Perhaps this is because an annual measure can include more months prior to frst birth.